+++ Shadow.Adapter/Adapter.fs
@@ -8,6 +8,7 @@ module Adapter =
   let DoResume() = Instance.DoResume null
   let VisitsClear() = Instance.Visits.Clear()
   let SamplesClear() = Instance.Samples.Clear()
+  let FlushAll() = Instance.FlushAll ProcessExit
 
   let internal prepareName name =
     if name